requirements.txt
# local package
-e .

# external requirements
coverage
flake8
python-dotenv>=0.5.1
numpy
pandas
huggingface
transformers
torch
matplotlib
datasets
ipykernel
openai
plotly
scipy
scikit-learn
faiss-cpu
typing
tqdm
langchain
fastapi
uvicorn